+++ Cqrs/Commands/Validators/CreateQuizCommandValidator.cs
@@ -1,13 +1,32 @@
-﻿using AspNetCoreVueStarter.Models;
+﻿using AspNetCoreVueStarter.Data.Repositories.Interfaces;
+using AspNetCoreVueStarter.Models;
 using FluentValidation;
+using Microsoft.EntityFrameworkCore;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace AspNetCoreVueStarter.Cqrs.Commands.Validators
 {
     public class CreateQuizCommandValidator : AbstractValidator<CreateQuizCommand>
     {
-        public CreateQuizCommandValidator(IValidator<Quiz> quizValidator)
+        private readonly IQuizRepository _quizRepository;
+
+        public CreateQuizCommandValidator(IValidator<Quiz> quizValidator, IQuizRepository quizRepository)
         {
+            _quizRepository = quizRepository;
+
+            RuleFor(x => x.Name)
+                .MustAsync(HaveUniqueNameAsync)
+                .WithMessage("Quiz name must be unique.");
+
             RuleFor(x => x).SetValidator(quizValidator);
+        }
+
+        private async Task<bool> HaveUniqueNameAsync(string name, CancellationToken cancellationToken)
+        {
+            var quizExists = await _quizRepository.GetMany().AnyAsync(x => x.Name == name, cancellationToken);
+
+            return !quizExists;
         }
     }
 }
